more completion queues and threads
cpu% distribution between handleRpcs threads are not completly even.
Requesting a simple endpoint foo 100k times with 8 "requester" threads
and 12 worker threads on the server, will cause the CPU% distribution to
be 80% on waiting for a new request(when we are bombarding with requests
without stop) and 20% of real work.

Maybe there is a way change the polling behaviour.